Refactor class detail logic and navigation loading
This commit introduces several changes to improve the user experience and code quality. The useParams hook is now used in UpdatedDetailedComp to access the classId parameter directly. The logic for displaying homework details has been updated to handle cases where the classId does not match the newest class. Additionally, the VideoCard component has been refactored to conditionally render the "View All" button only when a toPage prop is provided.

In the DrawerComp component, the useNavigation hook is utilized to display a loading indicator when navigation is in progress. This provides a smoother transition between pages.

The ClassesDetailPage has been simplified by removing unnecessary state and useEffect logic, as the validation of the classId is no longer required. The loader function now also calculates the highest class number to determine the newest class, which is then passed to the UpdatedDetailedComp component.