Adjust video height in DetailComp based on prop
This commit introduces a new prop `detailComp` to the VideoPlayer component, which determines the height of the video player. When `detailComp` is true, the video player's height is set to 570px, otherwise, it defaults to 300px. This change allows for a more flexible video display in the DetailComp component. Additionally, unnecessary console logs and comments have been removed from the Exercises page and VideoPlayer component for cleaner code.